test(priceCalculator): 장바구니 계산기 테스트 완료
