--- src/mocks/mockResponse/initial-cart-items.json
@@ -1,40 +1,0 @@
-{
-  "content": [
-    {
-      "id": 4704,
-      "quantity": 2,
-      "product": {
-        "id": 11,
-        "name": "리복",
-        "price": 20000,
-        "imageUrl": "https://image.msscdn.net/images/goods_img/20221031/2909092/2909092_6_500.jpg",
-        "category": "fashion"
-      }
-    }
-  ],
-  "pageable": {
-    "sort": {
-      "sorted": false,
-      "unsorted": true,
-      "empty": true
-    },
-    "pageNumber": 0,
-    "pageSize": 20,
-    "offset": 0,
-    "paged": true,
-    "unpaged": false
-  },
-  "last": true,
-  "totalPages": 1,
-  "totalElements": 19,
-  "sort": {
-    "sorted": false,
-    "unsorted": true,
-    "empty": true
-  },
-  "first": true,
-  "number": 0,
-  "numberOfElements": 19,
-  "size": 20,
-  "empty": false
-}
